+++ grass-diary/src/components/Feed/Feed.tsx
@@ -19,7 +19,7 @@ interface IFeedProps {
 
 const Feed = ({ feed, isTop }: IFeedProps) => {
   const navigate = useNavigate();
-  const { memberId } = useUser();
+  const memberId = useUser();
   const { data: writer } = useWriterProfile(feed.memberId);
   const { modal, loginModal } = useModal();
