+++ grass-diary/src/components/Layout/Header.tsx
@@ -5,35 +5,11 @@ import { Profile } from '@components/index';
 import { useNavigate } from 'react-router-dom';
 import { useUser } from '@state/user/useUser';
 import { useModal } from '@state/modal/useModal';
-import { useEffect } from 'react';
-import { MODAL } from '@constants/message';
-import { semantic } from '@styles/semantic';
-import { INTERACTION } from '@styles/interaction';
 
 const Header = () => {
-  const { memberId, isError } = useUser();
-  const { modal, loginModal } = useModal();
+  const memberId = useUser();
+  const { loginModal } = useModal();
   const navigate = useNavigate();
-
-  useEffect(() => {
-    const manualLogout = localStorage.getItem('manualLogout');
-    if (isError && manualLogout === null) {
-      const setting = {
-        title: MODAL.login_expiration.title,
-        content: MODAL.login_expiration.content,
-      };
-
-      const button1 = {
-        active: true,
-        text: MODAL.confirm,
-        color: semantic.light.accent.solid.hero,
-        interaction: INTERACTION.accent.subtle(),
-      };
-
-      localStorage.removeItem('accessToken');
-      modal(setting, button1);
-    }
-  }, [isError]);
 
   return (
     <S.Header>
