+++ grass-diary/src/services/index.ts
@@ -22,7 +22,6 @@ API.interceptors.response.use(
   error => {
     if (error.response && error.response.status === 401) {
       localStorage.removeItem('accessToken');
-      window.location.href = '/';
     }
 
     return Promise.reject(error);
