+++ grass-diary/src/state/user/useUser.ts
@@ -5,8 +5,11 @@ import { useQuery } from '@tanstack/react-query';
 import API from '@services/index';
 import { END_POINT } from '@constants/api';
 import { AxiosError } from 'axios';
-import { useError } from '@hooks/useError';
 import { useMemberId, useSetMemberId } from './UserStore';
+import { MODAL } from '@constants/message';
+import { semantic } from '@styles/semantic';
+import { INTERACTION } from '@styles/interaction';
+import { useModal } from '@state/modal/useModal';
 
 const fetchAxios = async () => {
   const res = await API.get(END_POINT.member_info);
@@ -17,6 +20,7 @@ export const useUser = () => {
   const { isAuthenticated } = useAuth();
   const memberId = useMemberId();
   const setMemberId = useSetMemberId();
+  const { modal } = useModal();
 
   const { data, isSuccess, isError, error } = useQuery<
     number,
@@ -32,11 +36,36 @@ export const useUser = () => {
     refetchIntervalInBackground: true,
   });
 
+  if (isError) {
+    const manualLogout = localStorage.getItem('manualLogout');
+    console.log('error', error);
+    if (manualLogout === null) {
+      console.log('error 조건문 내부', error);
+
+      const setting = {
+        title: MODAL.authentication_error.title,
+        content:
+          `${error.response?.data.description}\n` + '다시 로그인 해주세요',
+      };
+
+      const button1 = {
+        active: true,
+        text: MODAL.confirm,
+        color: semantic.light.accent.solid.hero,
+        interaction: INTERACTION.accent.subtle(),
+        clickHandler: () => (window.location.href = '/'),
+      };
+
+      localStorage.removeItem('accessToken');
+      modal(setting, button1);
+    }
+  }
+
   useEffect(() => {
     if (!isAuthenticated) setMemberId(0);
     else if (isError) setMemberId(0);
     else if (isSuccess) setMemberId(data);
   }, [isAuthenticated, isError, isSuccess]);
 
-  return { memberId, isError };
+  return memberId;
 };
